+++ docs/readme/examples_source/extractor/train_with_pml_advanced.md
@@ -7,9 +7,7 @@ import torch
 from tqdm import tqdm
 
 from oml.datasets.base import DatasetWithLabels
-from oml.losses.triplet import TripletLossWithMiner
-from oml.miners.inbatch_all_tri import AllTripletsMiner
-from oml.models.vit.vit import ViTExtractor
+from oml.models import ViTExtractor
 from oml.samplers.balance import BalanceSampler
 from oml.utils.download_mock_dataset import download_mock_dataset
