+++ docs/readme/examples_source/postprocessing/train_val.md
@@ -3,6 +3,7 @@
 <p>
 
 [comment]:postprocessor-start
+
 ```python
 from pprint import pprint
 
@@ -14,8 +15,7 @@ from oml.datasets.base import DatasetWithLabels, DatasetQueryGallery
 from oml.inference.flat import inference_on_dataframe
 from oml.metrics.embeddings import EmbeddingMetrics
 from oml.miners.pairs import PairsMiner
-from oml.models.meta.siamese import ConcatSiamese
-from oml.models.vit.vit import ViTExtractor
+from oml.models import ConcatSiamese, ViTExtractor
 from oml.retrieval.postprocessors.pairwise import PairwiseImagesPostprocessor
 from oml.samplers.balance import BalanceSampler
 from oml.transforms.images.torchvision import get_normalisation_resize_torch
@@ -28,8 +28,8 @@ download_mock_dataset(dataset_root)
 extractor = ViTExtractor("vits16_dino", arch="vits16", normalise_features=False)
 transform = get_normalisation_resize_torch(im_size=64)
 
-embeddings_train, embeddings_val, df_train, df_val = \
-    inference_on_dataframe(dataset_root, "df.csv", extractor=extractor, transforms=transform)
+embeddings_train, embeddings_val, df_train, df_val =
+inference_on_dataframe(dataset_root, "df.csv", extractor=extractor, transforms=transform)
 
 # We are building Siamese model on top of existing weights and train it to recognize positive/negative pairs
 siamese = ConcatSiamese(extractor=extractor, mlp_hidden_dims=[100])
