+++ oml/models/resnet/extractor.py
@@ -11,7 +11,7 @@ from torch import nn
 from torchvision.models import resnet18, resnet34, resnet50, resnet101, resnet152
 
 from oml.interfaces.models import IExtractor
-from oml.models.pooling import GEM
+from oml.models.resnet.pooling import GEM
 from oml.models.utils import remove_prefix_from_state_dict
 from oml.transforms.images.albumentations import get_normalisation_albu
 from oml.utils.io import download_checkpoint
