+++ oml/models/vit_clip/external/model.py
