+++ tests/test_oml/test_models/test_models_creation.py
@@ -7,9 +7,9 @@ import torch
 from oml.interfaces.models import IExtractor, IPairwiseModel
 from oml.models.meta.projection import ExtractorWithMLP
 from oml.models.meta.siamese import ConcatSiamese
-from oml.models.resnet import ResnetExtractor
-from oml.models.vit.clip import ViTCLIPExtractor
-from oml.models.vit.vit import ViTExtractor
+from oml.models.resnet.extractor import ResnetExtractor
+from oml.models.vit_clip.extractor import ViTCLIPExtractor
+from oml.models.vit_dino.extractor import ViTExtractor
 
 SKIP_LARGE_CKPT = True
